Refactor seeders to use specific namespaces for better organization
