+++ Lume.Infrastructure/Extensions/ServiceCollectionExtensions.cs
@@ -4,6 +4,9 @@ using Lume.Infrastructure.Persistence;
 using Lume.Infrastructure.Persistence.Repositories;
 using Lume.Infrastructure.Persistence.Seeders;
 using Lume.Infrastructure.Persistence.Seeders.Profiles;
+using Lume.Infrastructure.Persistence.Seeders.ReservationSeeders;
+using Lume.Infrastructure.Persistence.Seeders.TableSeeders;
+using Lume.Infrastructure.Persistence.Seeders.UserSeeders;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
