--- Lume.Infrastructure/Persistence/Seeders/IReservationSeeder.cs
@@ -1,6 +1,0 @@
-namespace Lume.Infrastructure.Persistence.Seeders;
-
-public interface IReservationSeeder
-{
-    Task SeedAsync();
-}
