--- Lume.Infrastructure/Persistence/Seeders/ITableSeeder.cs
@@ -1,6 +1,0 @@
-namespace Lume.Infrastructure.Persistence.Seeders;
-
-public interface ITableSeeder
-{
-    Task SeedAsync();
-}
