--- Lume.Infrastructure/Persistence/Seeders/IUserSeeder.cs
@@ -1,6 +1,0 @@
-namespace Lume.Infrastructure.Persistence.Seeders;
-
-public interface IUserSeeder
-{
-    public Task SeedAsync();
-}
