+++ Lume.Infrastructure/Persistence/Seeders/ReservationSeeders/IReservationSeeder.cs
@@ -1,0 +1,6 @@
+namespace Lume.Infrastructure.Persistence.Seeders.ReservationSeeders;
+
+public interface IReservationSeeder
+{
+    Task SeedAsync();
+}
