+++ Lume.Infrastructure/Persistence/Seeders/TableSeeders/ITableSeeder.cs
@@ -1,0 +1,6 @@
+namespace Lume.Infrastructure.Persistence.Seeders.TableSeeders;
+
+public interface ITableSeeder
+{
+    Task SeedAsync();
+}
