+++ Lume.Infrastructure/Persistence/Seeders/UserSeeders/IUserSeeder.cs
@@ -1,0 +1,6 @@
+namespace Lume.Infrastructure.Persistence.Seeders.UserSeeders;
+
+public interface IUserSeeder
+{
+    public Task SeedAsync();
+}
